+++ tests/spec/jsAxios/schema.js
@@ -17,13 +17,17 @@ export var ContentType;
   ContentType["UrlEncoded"] = "application/x-www-form-urlencoded";
 })(ContentType || (ContentType = {}));
 export class HttpClient {
-  constructor({ securityWorker, ...axiosConfig } = {}) {
+  constructor({ securityWorker, secure, ...axiosConfig } = {}) {
     this.securityData = null;
     this.setSecurityData = (data) => {
       this.securityData = data;
     };
     this.request = async ({ secure, path, type, query, format = "json", body, ...params }) => {
-      const secureParams = (secure && this.securityWorker && (await this.securityWorker(this.securityData))) || {};
+      const secureParams =
+        ((typeof secure === "boolean" ? secure : this.secure) &&
+          this.securityWorker &&
+          (await this.securityWorker(this.securityData))) ||
+        {};
       const requestParams = this.mergeRequestParams(params, secureParams);
       return this.instance.request({
         ...requestParams,
@@ -38,6 +42,7 @@ export class HttpClient {
       });
     };
     this.instance = axios.create({ ...axiosConfig, baseURL: axiosConfig.baseURL || "https://api.github.com" });
+    this.secure = secure;
     this.securityWorker = securityWorker;
   }
   mergeRequestParams(params1, params2) {